Multiview Workflow
Multi-View documentation is now mostly complete. What needs to be decided is whether or not to move the individual parts to their respective areas. I started doing that (locally) and it can work, but maybe they are fine to be all in the same page.

Also I suspect some of my PNGs could be rescaled to save space