Fix T42914 fix remaining FixMe's
Took a little bit of work... Had to Google translate some text from the old wiki. Alot of the text needs to be revised and made less verbose. The BGE Physics needs alot of help however this is a big project for another day and we will want to see how it goes in the Blender 2.8 project. Alot of the text here is repeated in the different pages and needs to be re formated. I added a link as a code comment at the top of 	blender_docs\manual\game_engine\physics\object\dynamic.rst for future references. The same is true for Fluid docs. I also removed some images that were delted in recent commits.

git-svn-id: https://svn.blender.org/svnroot/bf-manual/trunk/blender_docs@924 c4de1f47-6596-e411-a384-0024e86c2797